cobra-cli: 1.3.0 -> 1.3.0-21-g1d43487 (fix zsh completions)
The last tagged release of cobra-cli still uses a very old and buggy
version of cobra, which can lead to the cli completions not loading on
zsh. This oversight has already been fixed in cobra-cli's master, but
was never released as a tagged version. (and likely never will, the
project seems mostly deserted)

This commit bumps the version of cobra-cli to the newest commit on
master (1d43487), which includes the bump to cobra in go.sum